separate message area to own component
